feat: refactor contract spec into own mod; to be moved upstream later
Also check if contract_id is a token if doesn't find the normal contract code
